move type normalization functions into sep file
Moves the cleanGoType and replacePkgName functions into a separate
`pkg/model/types.go` file and makes cleanGoType a normal function and
not a method on CRD.
